LayoutTests/http/tests/inspector/network/network-open-load-reopen.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>

function loadImagesAndReopenFrontend()
{
    var image = document.createElement("img");
    image.setAttribute("src", "resources/resource.php?type=image&size=500");
    image.onload = resourceLoaded;
    document.body.appendChild(image);

    var image2 = document.createElement("img");
    image2.onload = resourceLoaded;
    image2.setAttribute("src", "resources/resource.php?type=image&size=500&gzip=true");
    document.body.appendChild(image2);
}

var loadedResourceCount = 0;
function resourceLoaded()
{
    if (++loadedResourceCount === 2)
        reopenFrontend();
}

function test()
{
    try {
        InspectorTest.evaluateInPage("frontendReopeningCount", frontendReopeningCountReceiver);
    } catch (e) {
        InspectorTest.disableBackgroundEventCollection();
        throw e;
    }

    function frontendReopeningCountReceiver(stepNumber) {
        if (stepNumber._description === "0") {
            InspectorTest.enableBackgroundEventCollection();
            InspectorTest.evaluateInPage("loadImagesAndReopenFrontend()");
        } else {
            InspectorTest.addResult("Reopened frontend " + stepNumber._description + " times");
            InspectorTest.dumpNetworkResources();
            // Switch off before completeTest.
            InspectorTest.disableBackgroundEventCollection();
            InspectorTest.completeTest();
        }
    }
}

</script>

</head>

<body onload="runTest()">
<p>
Test that when background event collection is enabled network panel correctly restores network log after the frontend reopening.
<a href="https://bugs.webkit.org/show_bug.cgi?id=58064">Bug 58064</a>
<a href="https://bugs.webkit.org/show_bug.cgi?id=58652">Bug 58652</a>
</p>
</body>
</html>
