LayoutTests/http/tests/inspector/network/network-open-load-reopen-expected.txt

Test that when background event collection is enabled network panel correctly restores network log after the frontend reopening. Bug 58064 Bug 58652


Reopened frontend 1 times
resources count = 2
http://127.0.0.1:8000/inspector/network/resources/resource.php?type=image&size=500
http://127.0.0.1:8000/inspector/network/resources/resource.php?type=image&size=500&gzip=true
